task-based image for opendreambox
first attempt, could not find any drawback so far.
split up in so many small task(file)s so we do not build unneeded
packages (e.g. pull madwifi on dm7025)...

note: the ui/dvbapi tasks are virtual and controlled through a
preferred_provider - this is fine for now but feel free to change this.